grainrad-poc/requirements.txt
anthropic>=0.18.0
google-genai>=0.2.0
pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
